Added cancel Product and some test cases

+++ com.capgemini.go/src/main/java/com/capgemini/go/service/OrderService.java
@@ -2,7 +2,6 @@ package com.capgemini.go.service;
 
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,6 +19,8 @@ import com.capgemini.go.dto.Cart;
 import com.capgemini.go.dto.OrderProductMap;
 import com.capgemini.go.dto.Orders;
 import com.capgemini.go.dto.Product;
+import com.capgemini.go.exceptions.OrderException;
+import com.capgemini.go.exceptions.ProductException;
 
 @Service
 @Transactional
@@ -36,6 +37,7 @@ public class OrderService implements IOrderService{
 	@Autowired
 	private ProductRepository productRepository;
 
+
 	/*
 	 * Function Name : createNewOrder 
 	 * Input Parameters :  Orders
@@ -44,7 +46,8 @@ public class OrderService implements IOrderService{
 	 */
 
 	@Override
-	public String createNewOrder(Orders orders) {
+	public String createNewOrder(Orders orders) throws OrderException {
+		System.out.println("in orders");
 		if(validateOrder(orders)) {
 			String orderId = UUID.randomUUID().toString();
 			orders.setOrderId(orderId);
@@ -54,30 +57,33 @@ public class OrderService implements IOrderService{
 			c.add(Calendar.DATE, 2);
 			orders.setOrderDispatchTime(c.getTime());
 			List<Cart> products = cartRepository.getProductsByUserId(orders.getUserId());
-			List<OrderProductMap> orderProductMapList = new ArrayList<>();
-			for(Cart cart : products) {
-				OrderProductMap map = new OrderProductMap();
-				Product product = productRepository.getProductByProductId(cart.getProductId());
-				map.setOrderId(orderId);
-				map.setProductId(cart.getProductId());
-				map.setProductStatus(1);
-				map.setGiftStatus(1);
-				map.setProducts(product);
-				orderProductMapRepository.save(map);
-				orderProductMapList.add(map);
-				//cartRepository.deleteByUserIdAndProductId(orders.getUserId(),cart.getProductId());
+			if(!products.isEmpty()) {
+				List<OrderProductMap> orderProductMapList = new ArrayList<>();
+				for(Cart cart : products) {
+					OrderProductMap map = new OrderProductMap();
+					Product product = productRepository.getProductByProductId(cart.getProductId());
+					map.setOrderId(orderId);
+					map.setProductId(cart.getProductId());
+					map.setProductStatus(0);
+					map.setGiftStatus(0);
+					map.setProduct(product);
+					orderProductMapRepository.save(map);
+					orderProductMapList.add(map);
+					//cartRepository.deleteByUserIdAndProductId(orders.getUserId(),cart.getProductId());
+				}
+				orders.setProducts(orderProductMapList);
+				ordersRepository.save(orders);
+				return "order placed successfully";
+				//return orderId;
 			}
-			orders.setProducts(orderProductMapList);
-			System.out.println(orders);
-			ordersRepository.save(orders);
-			//return "order placed successfully";
-			return orderId;
+			else {
+				throw new OrderException("valid userId needed or zero items in the cart...!!!");
+			}
 		}
 		else {
-			return "sorry, order was not placed";
+			throw new OrderException("incorrect UserId or AddressId...!!!");
 		}
 	}
-
 
 	/*
 	 * Function Name : findOrdersByUserId 
@@ -87,26 +93,29 @@ public class OrderService implements IOrderService{
 	 */
 
 	@Override
-	public List<Orders> findOrdersByUserId(String userId){
-		System.out.println(userId);
+	public List<Orders> findOrdersByUserId(String userId) throws OrderException{
 		List<Orders> orders = new ArrayList<>();
-		/*
-		 * Validating userId
-		 */
+
+		//Validating userId
 		if(userId == null || userId.isEmpty()) {
-			return Collections.emptyList();
+			throw new OrderException("valid userId needed..!!");
 		}
 		else {
 			orders = ordersRepository.findUserById(userId);
-			for(Orders order : orders) {
-				List<OrderProductMap> products = orderProductMapRepository.getOrderProductMapByOrderId(order.getOrderId());
-				for(OrderProductMap map : products) {
-					Product product = productRepository.getProductByProductId(map.getProductId());
-					map.setProducts(product);
-				}
-				order.setProducts(products);
+			if(orders.isEmpty()) {
+				throw new OrderException("valid userId needed..!!");
 			}
-			return orders;
+			else {
+				for(Orders order : orders) {
+					List<OrderProductMap> products = orderProductMapRepository.getOrderProductMapByOrderId(order.getOrderId());
+					for(OrderProductMap map : products) {
+						Product product = productRepository.getProductByProductId(map.getProductId());
+						map.setProduct(product);
+					}
+					order.setProducts(products);
+				}
+				return orders;
+			}
 		}
 	}
 
@@ -114,19 +123,62 @@ public class OrderService implements IOrderService{
 	 * Function Name : cancelOrder 
 	 * Input Parameters :  orderId
 	 * Return Type : String 
-	 * Description : to delete order
+	 * Description : to cancel order
 	 */
 
 	@Override
-	public String cancelOrder(String orderId) {
+	public String cancelOrder(String orderId) throws OrderException {
 		if (orderId == null || orderId.isEmpty()) {
-			return "sorry, order was not cancelled";
+			throw new OrderException("Valid orderId needed...!!");
 		}
 		else {
-			ordersRepository.deleteByOrderId(orderId);
-			orderProductMapRepository.deleteByOrderId(orderId);
-			return "order was deleted successfully";
-		} 
+			Orders order = ordersRepository.getOrderByOrderId(orderId);
+			if(order == null) {
+				throw new OrderException("valid orderId needed..!!");
+			}
+			else {
+				ordersRepository.deleteByOrderId(orderId);
+				orderProductMapRepository.deleteByOrderId(orderId);
+				return "order was deleted successfully";
+			} 
+		}
+	}
+
+	/* 
+	 * Function Name : removeProduct 
+	 * Input Parameters :  productId
+	 * Return Type : String 
+	 * Description : to cancel product which is not yet dispatched
+	 */
+
+	@Override
+	public String removeProduct(String productId) throws ProductException {
+		if(productId == null || productId.isEmpty()) {
+			throw new ProductException("valid productId needed...!!!");
+		}
+		else {
+			OrderProductMap map = orderProductMapRepository.getProductByProductId(productId);
+			List<OrderProductMap> mapList = orderProductMapRepository.findAll();
+			if(map == null) {
+				throw new ProductException("valid productId needed...!!!");
+			}
+			else {
+				if(map.getProductStatus() == 0) {
+					if(mapList.size() == 1) {
+						orderProductMapRepository.deleteProductByProductId(productId);
+						ordersRepository.deleteByOrderId(map.getOrderId());
+						return "product was removed successfully, order cancelled..!!";
+					}
+					else {
+						orderProductMapRepository.deleteProductByProductId(productId);
+						return "product was removed successfully..!!";
+					}
+				}
+				else {
+					throw new ProductException("product has already been dispatched and cannot be cancelled...!!!");
+				}
+			}
+		}
 	}
 
 	/* 
@@ -144,5 +196,24 @@ public class OrderService implements IOrderService{
 		}
 		else
 			return true;
+	}
+
+	/* 
+	 * Function Name : addItemToCart 
+	 * Input Parameters :  Cart
+	 * Return Type : List<Cart>
+	 * Description : to validate order
+	 */
+	@Override
+	public List<Cart> addItemToCart(Cart cart) {
+		Product product = productRepository.getProductByProductId(cart.getProductId());
+		cart.setProduct(product);
+		cartRepository.save(cart);
+		List<Cart> cartList = cartRepository.getCartListByUserId(cart.getUserId());
+		for(Cart c : cartList) {
+			Product p = productRepository.getProductByProductId(c.getProductId());
+			c.setProduct(p);
+		}
+		return cartList;
 	}
 }
